+++ tests/UploadControllerTest.cs
@@ -7,6 +7,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
 using System.Net;
+using System.Text;
 using static BDMS.Helpers;
 using File = System.IO.File;
 
@@ -66,9 +67,9 @@ public class UploadControllerTest
         byte[] fileBytes = File.ReadAllBytes(csvFile);
         using var stream = new MemoryStream(fileBytes);
 
-        var file = new FormFile(stream, 0, fileBytes.Length, csvFile, "text/csv");
+        var boreholeCsvFile = new FormFile(stream, 0, fileBytes.Length, null, csvFile);
 
-        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, file);
+        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, boreholesFile: boreholeCsvFile, pdfAttachments: null);
 
         Assert.IsInstanceOfType(response.Result, typeof(OkObjectResult));
         OkObjectResult okResult = (OkObjectResult)response.Result!;
@@ -112,9 +113,9 @@ public class UploadControllerTest
         byte[] fileBytes = File.ReadAllBytes(csvFile);
         using var stream = new MemoryStream(fileBytes);
 
-        var file = new FormFile(stream, 0, fileBytes.Length, csvFile, "text/csv");
+        var boreholeCsvFile = new FormFile(stream, 0, fileBytes.Length, null, csvFile);
 
-        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, file);
+        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, boreholesFile: boreholeCsvFile, pdfAttachments: null);
 
         Assert.IsInstanceOfType(response.Result, typeof(OkObjectResult));
         OkObjectResult okResult = (OkObjectResult)response.Result!;
@@ -156,9 +157,9 @@ public class UploadControllerTest
         byte[] fileBytes = File.ReadAllBytes(csvFile);
         using var stream = new MemoryStream(fileBytes);
 
-        var file = new FormFile(stream, 0, fileBytes.Length, csvFile, "text/csv");
+        var boreholeCsvFile = new FormFile(stream, 0, fileBytes.Length, null, csvFile);
 
-        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, file);
+        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, boreholesFile: boreholeCsvFile, pdfAttachments: null);
 
         Assert.IsInstanceOfType(response.Result, typeof(OkObjectResult));
         OkObjectResult okResult = (OkObjectResult)response.Result!;
@@ -180,9 +181,9 @@ public class UploadControllerTest
         byte[] fileBytes = File.ReadAllBytes(csvFile);
         using var stream = new MemoryStream(fileBytes);
 
-        var file = new FormFile(stream, 0, fileBytes.Length, csvFile, "text/csv");
+        var boreholeCsvFile = new FormFile(stream, 0, fileBytes.Length, null, csvFile);
 
-        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, file);
+        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, boreholesFile: boreholeCsvFile, pdfAttachments: null);
 
         Assert.IsInstanceOfType(response.Result, typeof(ObjectResult));
         ObjectResult result = (ObjectResult)response.Result!;
@@ -207,9 +208,9 @@ public class UploadControllerTest
         byte[] fileBytes = File.ReadAllBytes(csvFile);
         using var stream = new MemoryStream(fileBytes);
 
-        var file = new FormFile(stream, 0, fileBytes.Length, csvFile, "text/csv");
+        var boreholeCsvFile = new FormFile(stream, 0, fileBytes.Length, null, csvFile);
 
-        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, file);
+        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, boreholesFile: boreholeCsvFile, pdfAttachments: null);
 
         Assert.IsInstanceOfType(response.Result, typeof(OkObjectResult));
         OkObjectResult okResult = (OkObjectResult)response.Result!;
@@ -241,9 +242,9 @@ public class UploadControllerTest
         byte[] fileBytes = File.ReadAllBytes(csvFile);
         using var stream = new MemoryStream(fileBytes);
 
-        var file = new FormFile(stream, 0, fileBytes.Length, csvFile, "text/csv");
+        var boreholeCsvFile = new FormFile(stream, 0, fileBytes.Length, null, csvFile);
 
-        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, file);
+        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, boreholesFile: boreholeCsvFile, pdfAttachments: null);
 
         Assert.IsInstanceOfType(response.Result, typeof(OkObjectResult));
         OkObjectResult okResult = (OkObjectResult)response.Result!;
@@ -275,9 +276,9 @@ public class UploadControllerTest
         byte[] fileBytes = File.ReadAllBytes(csvFile);
         using var stream = new MemoryStream(fileBytes);
 
-        var file = new FormFile(stream, 0, fileBytes.Length, csvFile, "text/csv");
+        var boreholeCsvFile = new FormFile(stream, 0, fileBytes.Length, null, csvFile);
 
-        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, file);
+        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, boreholesFile: boreholeCsvFile, pdfAttachments: null);
 
         Assert.IsInstanceOfType(response.Result, typeof(OkObjectResult));
         OkObjectResult okResult = (OkObjectResult)response.Result!;
@@ -299,23 +300,124 @@ public class UploadControllerTest
     [TestMethod]
     public async Task UploadEmptyFileShouldReturnError()
     {
-        var file = new FormFile(null, 0, 0, "nonexistant_file.csv", "text/csv");
+        var boreholeCsvFile = new FormFile(null, 0, 0, null, "nonexistant_file.csv");
 
-        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, file);
+        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, boreholesFile: boreholeCsvFile, pdfAttachments: null);
 
         Assert.IsInstanceOfType(response.Result, typeof(BadRequestObjectResult));
         BadRequestObjectResult badRequestResult = (BadRequestObjectResult)response.Result!;
-        Assert.AreEqual("No file uploaded.", badRequestResult.Value);
+        Assert.AreEqual("No borehole csv file uploaded.", badRequestResult.Value);
+    }
+
+    [TestMethod]
+    public async Task UploadInvalidFileTypeBoreholeCsvShouldReturnError()
+    {
+        string fileContent = "This is the content of the file.";
+        byte[] fileBytes = Encoding.UTF8.GetBytes(fileContent);
+        var invalidFielTypeBoreholeFile = new FormFile(new MemoryStream(fileBytes), 0, fileBytes.Length, null, "invalid_file_type.txt");
+
+        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, boreholesFile: invalidFielTypeBoreholeFile, pdfAttachments: null);
+
+        Assert.IsInstanceOfType(response.Result, typeof(BadRequestObjectResult));
+        BadRequestObjectResult badRequestResult = (BadRequestObjectResult)response.Result!;
+        Assert.AreEqual("Invalid file type for borehole csv.", badRequestResult.Value);
+    }
+
+    [TestMethod]
+    public async Task UploadBoreholeCsvFileWithNotLinkedPdfsShouldCreateBorehole()
+    {
+        httpClientFactoryMock
+           .Setup(cf => cf.CreateClient(It.IsAny<string>()))
+           .Returns(() => new HttpClient())
+           .Verifiable();
+
+        string fileContent = "original_name;location_x;location_y\r\nFrank Place;2000000;1000000";
+        byte[] fileBytes = Encoding.UTF8.GetBytes(fileContent);
+        var boreholeCsvFile = new FormFile(new MemoryStream(fileBytes), 0, fileBytes.Length, null, "invalid_file_type.csv");
+
+        var firstPdfFile = "borehole_attachment_1.pdf";
+        fileBytes = File.ReadAllBytes(firstPdfFile);
+        var firstPdfFormFile = new FormFile(new MemoryStream(File.ReadAllBytes(firstPdfFile)), 0, fileBytes.Length, null, "first_pdf.pdf");
+
+        var secondPdfFile = "borehole_attachment_2.pdf";
+        fileBytes = File.ReadAllBytes(secondPdfFile);
+        var secondPdfFormFile = new FormFile(new MemoryStream(File.ReadAllBytes(secondPdfFile)), 0, fileBytes.Length, null, "second_pdf.pdf");
+
+        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1,
+                                                                      boreholesFile: boreholeCsvFile,
+                                                                      pdfAttachments: new List<IFormFile>() { firstPdfFormFile, secondPdfFormFile });
+
+        Assert.IsInstanceOfType(response.Result, typeof(OkObjectResult));
+        OkObjectResult okResult = (OkObjectResult)response.Result!;
+        Assert.AreEqual(1, okResult.Value);
+    }
+
+    [TestMethod]
+    public async Task UploadInvalidFileTypePdfAttachmentShouldReturnError()
+    {
+        string fileContent = "This is the content of the file."; // set file content
+        byte[] fileBytes = Encoding.UTF8.GetBytes(fileContent);
+        var boreholeCsvFile = new FormFile(new MemoryStream(fileBytes), 0, fileBytes.Length, null, "boreholes.csv");
+
+        var firstPdfFile = "borehole_attachment_1.pdf";
+        fileBytes = File.ReadAllBytes(firstPdfFile);
+        var firstPdfFormFile = new FormFile(new MemoryStream(fileBytes), 0, fileBytes.Length, null, "first_pdf.pdf");
+
+        var secondPdfFile = "borehole_attachment_with_wrong_extension.txt";
+        fileBytes = File.ReadAllBytes(secondPdfFile);
+        var secondPdfFormFile = new FormFile(new MemoryStream(fileBytes), 0, fileBytes.Length, null, "borehole_attachment_with_wrong_extension.txt");
+
+        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1,
+                                                                      boreholesFile: boreholeCsvFile,
+                                                                      pdfAttachments: new List<IFormFile>() { firstPdfFormFile, secondPdfFormFile });
+
+        Assert.IsInstanceOfType(response.Result, typeof(BadRequestObjectResult));
+        BadRequestObjectResult badRequestResult = (BadRequestObjectResult)response.Result!;
+        Assert.AreEqual("Invalid file type for pdf attachment.", badRequestResult.Value);
+    }
+
+    [TestMethod]
+    public void IsCorrectFileType()
+    {
+        string fileContent = "This is the content of the file."; // set file content
+        byte[] fileBytes = Encoding.UTF8.GetBytes(fileContent);
+        var validCsvFile = new FormFile(new MemoryStream(fileBytes), 0, fileBytes.Length, null, "boreholes.csv");
+        Assert.AreEqual(true, FileTypeChecker.IsCsv(validCsvFile));
+
+        var emptyCsvFile = new FormFile(null, 0, fileBytes.Length, null, "boreholes.csv");
+        Assert.AreEqual(true, FileTypeChecker.IsCsv(emptyCsvFile));
+
+        var invalidCsvFile = new FormFile(new MemoryStream(fileBytes), 0, fileBytes.Length, null, "boreholes.txt");
+        Assert.AreEqual(false, FileTypeChecker.IsCsv(invalidCsvFile));
+
+        fileBytes = File.ReadAllBytes("borehole_attachment_1.pdf");
+        var validPdfFormFile = new FormFile(new MemoryStream(fileBytes), 0, fileBytes.Length, null, "attachment.pdf");
+        Assert.AreEqual(true, FileTypeChecker.IsPdf(validPdfFormFile));
+
+        var zeroBytesFormFile = new FormFile(null, 0, 0, null, "attachment.pdf");
+        Assert.AreEqual(false, FileTypeChecker.IsPdf(zeroBytesFormFile));
+
+        fileBytes = File.ReadAllBytes("borehole_attachment_with_wrong_extension.txt");
+        var invalidExtensionPdfFormFile = new FormFile(new MemoryStream(fileBytes), 0, fileBytes.Length, null, "borehole_attachment_with_wrong_extension.txt");
+        Assert.AreEqual(false, FileTypeChecker.IsPdf(invalidExtensionPdfFormFile));
+
+        fileBytes = Encoding.UTF8.GetBytes("This is not a PDF.");
+        var invalidHeaderBytesPdfFormFile = new FormFile(new MemoryStream(fileBytes), 0, fileBytes.Length, null, "attachment.pdf");
+        Assert.AreEqual(false, FileTypeChecker.IsPdf(invalidHeaderBytesPdfFormFile));
+
+        fileBytes = File.ReadAllBytes("empty_attachment.pdf");
+        var emptyPdfFormFile = new FormFile(new MemoryStream(fileBytes), 0, fileBytes.Length, null, "attachment.pdf");
+        Assert.AreEqual(true, FileTypeChecker.IsPdf(emptyPdfFormFile));
     }
 
     [TestMethod]
     public async Task UploadNoFileShouldReturnError()
     {
-        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, null);
+        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, boreholesFile: null, pdfAttachments: null);
 
         Assert.IsInstanceOfType(response.Result, typeof(BadRequestObjectResult));
         BadRequestObjectResult badRequestResult = (BadRequestObjectResult)response.Result!;
-        Assert.AreEqual("No file uploaded.", badRequestResult.Value);
+        Assert.AreEqual("No borehole csv file uploaded.", badRequestResult.Value);
     }
 
     [TestMethod]
@@ -326,9 +428,9 @@ public class UploadControllerTest
         byte[] fileBytes = File.ReadAllBytes(csvFile);
         using var stream = new MemoryStream(fileBytes);
 
-        var file = new FormFile(stream, 0, fileBytes.Length, csvFile, "text/csv");
+        var boreholeCsvFile = new FormFile(stream, 0, fileBytes.Length, null, csvFile);
 
-        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, file);
+        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, boreholesFile: boreholeCsvFile, pdfAttachments: null);
 
         Assert.IsInstanceOfType(response.Result, typeof(OkObjectResult));
         OkObjectResult okResult = (OkObjectResult)response.Result!;
@@ -343,9 +445,9 @@ public class UploadControllerTest
         byte[] fileBytes = File.ReadAllBytes(csvFile);
         using var stream = new MemoryStream(fileBytes);
 
-        var file = new FormFile(stream, 0, fileBytes.Length, csvFile, "text/csv");
+        var boreholeCsvFile = new FormFile(stream, 0, fileBytes.Length, null, csvFile);
 
-        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, file);
+        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, boreholesFile: boreholeCsvFile, pdfAttachments: null);
 
         Assert.IsInstanceOfType(response.Result, typeof(ObjectResult));
         ObjectResult result = (ObjectResult)response.Result!;
@@ -372,9 +474,9 @@ public class UploadControllerTest
         byte[] fileBytes = File.ReadAllBytes(csvFile);
         using var stream = new MemoryStream(fileBytes);
 
-        var file = new FormFile(stream, 0, fileBytes.Length, csvFile, "text/csv");
+        var boreholeCsvFile = new FormFile(stream, 0, fileBytes.Length, null, csvFile);
 
-        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, file);
+        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, boreholesFile: boreholeCsvFile, pdfAttachments: null);
 
         Assert.IsInstanceOfType(response.Result, typeof(ObjectResult));
         ObjectResult result = (ObjectResult)response.Result!;
@@ -394,7 +496,7 @@ public class UploadControllerTest
         byte[] fileBytes = File.ReadAllBytes(csvFile);
         using var stream = new MemoryStream(fileBytes);
 
-        var file = new FormFile(stream, 0, fileBytes.Length, csvFile, "text/csv");
+        var file = new FormFile(stream, 0, fileBytes.Length, null, csvFile);
 
         ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, file);
 
@@ -442,7 +544,7 @@ public class UploadControllerTest
         byte[] fileBytes = File.ReadAllBytes(csvFile);
         using var stream = new MemoryStream(fileBytes);
 
-        var file = new FormFile(stream, 0, fileBytes.Length, csvFile, "text/csv");
+        var file = new FormFile(stream, 0, fileBytes.Length, null, csvFile);
 
         ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, file);
 
@@ -490,9 +592,9 @@ public class UploadControllerTest
         byte[] fileBytes = File.ReadAllBytes(csvFile);
         using var stream = new MemoryStream(fileBytes);
 
-        var file = new FormFile(stream, 0, fileBytes.Length, csvFile, "text/csv");
+        var boreholeCsvFile = new FormFile(stream, 0, fileBytes.Length, null, csvFile);
 
-        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, file);
+        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, boreholesFile: boreholeCsvFile, pdfAttachments: null);
 
         Assert.IsInstanceOfType(response.Result, typeof(OkObjectResult));
         OkObjectResult okResult = (OkObjectResult)response.Result!;