+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Remove _clone_ and _delete_ icons in chronostratigraphy header.
 - Make csv import validation error modal content scrollable.
 - Replace csv import validation messages with more specific messages.
+- Update csv borehole upload endpoint to additionally handle list of PDF's.
 
 ### Fixed